Refactored Recommender
Recommender files changed and database structures improved
