--- recommender/recommend/settings_base.py
@@ -1,6 +1,0 @@
-import environ
-
-env = environ.Env()
-
-DB_URL = env("DATABASE_URL")
-
