+++ recommender/settings_base.py
